Mocking framework for external package calls in Go.
Added initial tests, and mocking support for most function types. 

Added the ability to to a dumb mock, which will return zero values for any returns of the original function. 
Added the ability to do specify the return values when calling the mock function.
Added the ability to specify another function to be called instead of the original.

All calls will be recorded and you can query the call history.